services/accounts/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.4
python-multipart==0.0.20
httpx==0.28.1
redis==5.2.1
pyotp==2.9.0
alembic==1.14.1
